Initialize repository
This repository contains a project implementing
the basic requirements of Signafire's take-home
code exercise.

This exercise requires the analysis of a csv log
file and a set of command-line inputs denoting
timestamps. The amount and identity of rows in
the log including each time provided by the user
are printed to the console. Additionally, some
descriptive statistics are generated.

Additionally, a foundation for documentation and
tests is included.